# internal/adapters/cfg/restapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Mockwave Admin API
  version: "0.2.0"
  description: >
    HTTP API for managing rules, simulations, metrics, and unmatched requests
    in a Mockwave instance. All endpoints are served on the admin port (default 9090).
servers:
  - url: http://localhost:9090
    description: Local instance
paths:
...
  /api/rules:
    get:
      summary: List all rules
      operationId: listRules
      responses:
        "200":
          description: Array of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
    post:
      summary: Create a rule
      operationId: createRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Rule"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/rules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a rule by ID
      operationId: getRule
      responses:
        "200":
          description: Rule found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update (replace) a rule
      operationId: updateRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Rule"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      summary: Delete a rule
      operationId: deleteRule
      responses:
        "204":
          description: Rule deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/simulations:
    get:
      summary: List all simulations
      operationId: listSimulations
      responses:
        "200":
          description: Array of simulations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Simulation"
    post:
      summary: Create a simulation
      operationId: createSimulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Simulation"
      responses:
        "201":
          description: Simulation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/simulations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a simulation by ID
      operationId: getSimulation
      responses:
        "200":
          description: Simulation found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update (replace) a simulation
      operationId: updateSimulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Simulation"
      responses:
        "200":
          description: Simulation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Simulation"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      summary: Delete a simulation
      operationId: deleteSimulation
      responses:
        "204":
          description: Simulation deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/export:
    get:
      summary: Export rules and their simulations as a Config file
      operationId: exportConfig
      parameters:
        - name: rules
          in: query
          required: false
          description: >-
            Comma-separated list of rule IDs to export. When omitted all rules
            are exported. Returns 404 if any supplied ID does not exist.
          schema:
            type: string
            example: "rule-1,rule-2"
      responses:
        "200":
          description: >-
            Config JSON ready to import. Response includes
            Content-Disposition: attachment; filename="mockwave-export.json".
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/import/preview:
    post:
      summary: Dry-run import — detect conflicts without writing
      operationId: importPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Config"
      responses:
        "200":
          description: Conflict report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreviewReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/import:
    post:
      summary: Commit an import — write rules and simulations to the store
      operationId: importConfig
      parameters:
        - name: override
          in: query
          required: false
          description: >-
            Comma-separated list of incoming rule IDs that should replace their
            conflicting counterpart in the store. Conflicting rules whose ID is
            not listed here are skipped.
          schema:
            type: string
            example: "rule-1,rule-2"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Config"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/faults:
    get:
      summary: List all fault profiles
      operationId: listFaultProfiles
      responses:
        "200":
          description: Array of fault profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaultProfile"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
    post:
      summary: Create a fault profile
      operationId: createFaultProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaultProfile"
      responses:
        "201":
          description: Fault profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaultProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /api/faults/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a fault profile by ID
      operationId: getFaultProfile
      responses:
        "200":
          description: Fault profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaultProfile"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
    put:
      summary: Update (replace) a fault profile
      operationId: updateFaultProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaultProfile"
      responses:
        "200":
          description: Fault profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaultProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
    delete:
      summary: Delete a fault profile
      operationId: deleteFaultProfile
      responses:
        "204":
          description: Fault profile deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /api/chaos/halt:
    post:
      summary: Globally halt fault injection (kill switch)
      operationId: chaosHalt
      responses:
        "204":
          description: Fault injection halted
        "501":
          $ref: "#/components/responses/NotImplemented"
  /api/chaos/resume:
    post:
      summary: Resume fault injection after a halt
      operationId: chaosResume
      responses:
        "204":
          description: Fault injection resumed
        "501":
          $ref: "#/components/responses/NotImplemented"
  /api/chaos/status:
    get:
      summary: Current kill-switch status
      operationId: chaosStatus
      responses:
        "200":
          description: Kill-switch status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChaosStatus"
        "501":
          $ref: "#/components/responses/NotImplemented"
  /api/scenarios:
    get:
      summary: List chaos scenarios
      operationId: listScenarios
      responses:
        "200":
          description: Scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
        "501":
          $ref: "#/components/responses/NotImplemented"
    post:
      summary: Create a scenario
      operationId: createScenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Scenario"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "501":
          $ref: "#/components/responses/NotImplemented"
...
components:
  schemas:
...
    Rule:
      type: object
      required: [id, match, buckets]
      properties:
        id:
          type: string
        name:
          type: string
        match:
          $ref: "#/components/schemas/MatchCriteria"
        buckets:
          type: array
          description: >-
            Weighted traffic buckets. Bucket weights must sum to exactly 100.
          items:
            $ref: "#/components/schemas/WeightedBucket"
...
    Simulation:
      type: object
      required: [id, protocol]
      properties:
        id:
          type: string
        protocol:
          type: string
          enum: [http, graphql, soap, grpc]
        response:
          $ref: "#/components/schemas/HTTPResponse"
        script:
          type: string
          description: Optional JavaScript (goja) run on every matched request
        soap_envelope:
          type: string
        grpc_message:
          type: string
        grpc_status:
          type: integer
    Config:
      type: object
      description: >-
        Top-level import/export envelope. Contains a list of rules and the
        simulations they reference.
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/Rule"
        simulations:
          type: array
          items:
            $ref: "#/components/schemas/Simulation"
...
    ImportPreviewReport:
      type: object
      required: [importable, conflicts]
      properties:
        importable:
          type: integer
          description: Number of rules that can be imported without conflict.
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/ImportConflict"
    ImportReport:
      type: object
      required: [imported, skipped, overridden]
      properties:
        imported:
          type: array
          description: IDs of rules that were newly written to the store.
          items:
            type: string
        skipped:
          type: array
          description: IDs of rules that were skipped due to unresolved conflicts.
          items:
            type: string
        overridden:
          type: array
          description: IDs of rules that replaced an existing conflicting rule.
          items:
            type: string
...
  responses:
    BadRequest:
      description: Invalid JSON or request body
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Forbidden:
      description: >-
        Feature is disabled. Import/export is not available when the json-file
        store is in use.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    MethodNotAllowed:
      description: HTTP method not allowed for this endpoint
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string